+++ src/compas_cem/diagrams/__init__.py
@@ -1,28 +1,11 @@
 """
-compas_cem.diagrams
-****************************
-
-.. currentmodule:: compas_cem.diagrams
-
-
-Diagrams
-========
-
-.. autosummary::
-    :toctree: generated/
-    :nosignatures:
-
-    TopologyDiagram
-    FormDiagram
+Topology and form diagrams.
 """
 
 from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-
-# from .<module> import *
-from .mixins import *  # noqa F403
 from .diagram import *  # noqa F403
 from .topology import *  # noqa F403
 from .form import *  # noqa F403