+++ src/compas_cem/ghpython/components/CompasCem_OriginNodesMove/code.py
@@ -4,7 +4,7 @@ Move the origin nodes of a topology diagram to a new location.
 
 from ghpythonlib.componentbase import executingcomponent as component
 
-from compas.utilities import iterable_like
+from compas.itertools import iterable_like
 from compas_rhino.geometry import RhinoPoint